Lightweight proxy For IAM
Added a lightweight proxy for IAM, running with the settings.json file this should run when the server is a private server and needs IAM premessions, this script should run and handle all Gcloud Auth by obtaining ODIC tokens, authorizing HTTP Post to the cloud run service, then it will collect a JSON RPC payload from the client and send it over to as an HTTP request. I am having a little trouble with the last part, as i get a 406 error sometimes however.